feat(damage/respawn)

+++ Assets/PlayerWeapon.cs
@@ -6,6 +6,7 @@ public class PlayerWeapon: MonoBehaviour {
 	[SerializeField]
 	public string weaponName = "UBER-OMEGA-LEET-MLG-RAYGUN";
 
-	public float weaponDamage = 100f;
+	public int weaponDamage = 100;
+
 	public float weaponRange = 300f;
 }
